Add database

+++ src/main/java/ru/yandex/practicum/javafilmorate/controllers/FilmController.java
@@ -15,6 +15,7 @@ public class FilmController {
 
     private final FilmService filmService;
 
+
     public FilmController(FilmService filmService) {
         this.filmService = filmService;
     }
@@ -43,7 +44,6 @@ public class FilmController {
         filmService.deleteLikeFilmInStorage(id, userId);
     }
 
-
     @GetMapping("/{id}")
     public Film getFilm(@PathVariable long id) {
         log.info(String.format("Получен запрос 'GET /films/%d'", id));
@@ -57,8 +57,9 @@ public class FilmController {
     }
 
     @GetMapping("/popular")
-    public List<Film> getMostLikedFilms(@RequestParam(defaultValue = "10") long count) {
+    public List<Film> getMostLikedFilms(@RequestParam(defaultValue = "10") int count) {
         log.info(String.format("Получен запрос 'GET /films/popular?count=%d'", count));
         return filmService.getMostLikedFilmsFromStorage(count);
     }
+
 }
